# [utilities,plugins] speedup journal collection (v2)