@pmoravec 2024-12-18
I can confirm VERY evident speedup on a tested RHEL8 system with:

```
# journalctl --disk-usage 
Archived and active journals take up 560.0M in the file system.
#
```

where current upstream executes `python3 bin/sos report -o logs --batch --build` in 18s and this PR shrinks it to 3s.

Some test is failing - can you @champtar review it or do you need some help? I might get to the failure later today  /tomorrow.