@champtar 2024-12-18
I'm off until the 27th so no rush for the review, and I was thinking about even implementing 'head' in Python