@champtar 2025-01-30
Not in the PR, but I was thinking about making TailReader work with to_file at some point, as I think we don't use the 'output' that often