@pmoravec 2025-01-30
So far the PR survived various my corner-case tests I invented. I shall review the code tomorrow..