@pmoravec 2025-01-31
All up to one nitpick is good. I confirm great performance gain (and no visible performance penalty for journals having 100MB where the new approach additionally has to `tac` the data). Several corner cases checked, no issue found. Code seems good.

I like the idea, but would welcome another profound review of others as well - since we touch a frequently used code.